API/requirements.txt
# Core ML (Tus versiones originales)
xgboost==2.0.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4

# Conexión a Base de Datos
pyodbc

# Requerido por Azure ML para el despliegue
azureml-defaults
joblib

# Para probar la API localmente
requests